Create systems submodule
