+++ src/chromatix/__init__.py
@@ -3,4 +3,4 @@ __version__ = "0.1"
 
 
 from .field import Field
-from .optical_system import OpticalSystem, Microscope
+from .systems import OpticalSystem
