+++ src/chromatix/systems/__init__.py
@@ -1,0 +1,2 @@
+from .optical_system import OpticalSystem
+from .microscopes import *
